Download tables from an Opal project as a dossier

Downloads a dossier or dataset from an Opal environment to the local environment. This is a wrapper function foropalr::opal.table_get() and opalr::opal.table_dictionary_get().

opal_tables_pull(
  opal,
  project,
  table_list = NULL,
  content = c("dataset", "data_dict"),
  keep_as_dossier = TRUE,
  .remove_id = FALSE
)

Arguments

opal

Opal login attributes.

project

A character string specifying the Opal project name.

table_list

A vector character string specifying Opal tables name.

content

A vector of character string which indicates if the function returns a dataset, or data dictionary. Default is 'dataset'.

keep_as_dossier

whether to return a dossier or a dataset if there is only one table. TRUE by default, if FALSE returns dataset.

.remove_id

whether to return the id column created in Opal or not. TRUE by default.

Value

Objects (tibbles and list of tibbles) representing tables and their respective data dictionary.

Details

Opal is the OBiBa application for data management, and Opal environments can be integrated with RStudio environments using the package opalr. The user must have adequate credentials to interact with a specific Opal environment. Some function errors produced may be associated with the handler or Opal read/write permissions.

See also

Examples

# \donttest{

library(opalr)
opal <- 
 opal.login('administrator','password',
   url ='https://opal-demo.obiba.org/')
   
# use DEMO_files provided by the package
library(madshapR)
library(stringr)

dossier <- 
  DEMO_files[str_detect(names(DEMO_files),"dataset_MELBOURNE")]

tempdir <- basename(tempdir())
try(opal_project_create(opal, tempdir))
#> The project RtmpmQTIRq already exists in Opal and will not be created here.

try(
  opal_tables_push(
  opal, dossier,project_name = tempdir, .force = TRUE, .overwrite = TRUE))
#> Verification of input format.
#> Verification of input format done.
#> Warning: Tables overwritten, or specify .overwrite = FALSE
#> 
#> The table: dataset_MELBOURNE_1 has been successfuly uploaded to Opal in RtmpmQTIRq
#> 
#> The table: dataset_MELBOURNE_2 has been successfuly uploaded to Opal in RtmpmQTIRq
  
###### Example pull a table from a project.
try(
  opal_tables_pull(
  opal,project = tempdir,table_list = 'dataset_MELBOURNE_1'))
#> Download of: RtmpmQTIRq / dataset_MELBOURNE_1 (1/1)
#> Error : Export of "RtmpmQTIRq.dataset_MELBOURNE_1" ended with status: FAILED

# }