Built-in tibble of allowed valueType values

Provides a built-in tibble showing the list of allowed Opal valueType values and their corresponding R data types. This tibble is mainly used for internal processes and programming.

valueType_list

Format

tibble

A data frame with 12 rows and 6 columns:

valueType

data type as described in Opal

typeof

data type provided by base::typeof

class

data class provided by base::class

call

function to transpose object according base::do.call function

toValueType

ensemble data type as described in Opal

toTypeof

ensemble data type provided by base::typeof

genericType

ensemble data type which valueType belongs

...

Details

The valueType is a property of a variable and is required in certain functions to determine the handling of the variables. The valueType refers to the OBiBa-internal type of a variable. It is specified in a data dictionary in a column valueType and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'). The full list of OBiBa valueType possibilities and their correspondence with R data types are available using valueType_list.

See also

Examples

{

valueType_list

}
#> # A tibble: 12 × 7
#>    valueType  typeof    class          call     toValueType toTypeof genericType
#>    <chr>      <chr>     <chr>          <chr>    <chr>       <chr>    <chr>      
#>  1 NA         character character      as.char… text        charact… character  
#>  2 text       character character      as.char… text        charact… character  
#>  3 integer    integer   integer        as.inte… integer     integer  numeric    
#>  4 decimal    double    numeric        as.nume… decimal     double   numeric    
#>  5 boolean    logical   logical        as_any_… boolean     logical  numeric    
#>  6 locale     character character      as.char… text        charact… character  
#>  7 datetime   double    POSIXct,POSIXt as.char… text        charact… character  
#>  8 date       double    Date           as_any_… decimal     double   date       
#>  9 binary     character character      as.char… text        charact… character  
#> 10 point      character character      as.char… text        charact… character  
#> 11 linestring character character      as.char… text        charact… character  
#> 12 polygon    character character      as.char… text        charact… character